fix: comment 수정, 삭제 버그 수정

See merge request s08-ai-image-sub2/S08P22B301!101

+++ Backend/src/main/java/com/ssafy/project/api/service/CommentServiceImpl.java
@@ -42,6 +42,7 @@ public class CommentServiceImpl implements CommentService{
                    .id(comment.getId())
                    .userProfile(comment.getUser().getSocialAuth().getImageUrl())
                    .nickname(comment.getUser().getNickname())
+                   .email(comment.getUser().getEmail())
                    .content(comment.getContent())
                    .createdDate(comment.getCreatedDate())
                    .build());
@@ -60,11 +61,11 @@ public class CommentServiceImpl implements CommentService{
                         .id(comment.getId())
                         .userProfile(comment.getUser().getSocialAuth().getImageUrl())
                         .nickname(comment.getUser().getNickname())
+                        .email(comment.getUser().getEmail())
                         .content(comment.getContent())
                         .title(comment.getBoard().getTitle())
                         .createdDate(comment.getCreatedDate())
                         .build());
-
         return commentDTOS;
     }
 
@@ -85,13 +86,14 @@ public class CommentServiceImpl implements CommentService{
     @Override
     public void modifyComment(CommentModReqDTO commentModReqDTO) {
         Comment comment = commentRepository.findById(commentModReqDTO.getCommentId()).orElseThrow(() -> new CommonApiException(CommonErrorCode.COMMENT_NOT_FOUND));
-
+        if(authProvider.getUserIdFromPrincipal() != comment.getUser().getId()) throw new CommonApiException(CommonErrorCode.COMMENT_NOT_ALLOWED);
         comment.setContent(commentModReqDTO.getContent());
     }
 
     @Override
     public void removeComment(Long commentId) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new CommonApiException(CommonErrorCode.COMMENT_NOT_FOUND));
+        if(authProvider.getUserIdFromPrincipal() != comment.getUser().getId()) throw new CommonApiException(CommonErrorCode.COMMENT_NOT_ALLOWED);
         commentRepository.deleteById(commentId);
     }
 }
